refactor: dto 위치 수정 및 반영 #34
