+++ src/main/java/pingpong/backend/domain/notion/controller/NotionOauthController.java
@@ -7,9 +7,9 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import pingpong.backend.domain.member.Member;
-import pingpong.backend.domain.notion.dto.NotionConnectDatabaseRequest;
-import pingpong.backend.domain.notion.dto.NotionOAuthExchangeRequest;
-import pingpong.backend.domain.notion.dto.NotionOAuthExchangeResponse;
+import pingpong.backend.domain.notion.dto.request.NotionConnectDatabaseRequest;
+import pingpong.backend.domain.notion.dto.request.NotionOAuthExchangeRequest;
+import pingpong.backend.domain.notion.dto.response.NotionOAuthExchangeResponse;
 import pingpong.backend.domain.notion.service.NotionFacade;
 import pingpong.backend.global.annotation.CurrentMember;
 import pingpong.backend.global.response.result.SuccessResponse;
