+++ src/main/java/pingpong/backend/domain/notion/controller/NotionResourceController.java
@@ -6,8 +6,8 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import pingpong.backend.domain.member.Member;
-import pingpong.backend.domain.notion.dto.NotionCreatePageRequest;
-import pingpong.backend.domain.notion.dto.NotionPageUpdateRequest;
+import pingpong.backend.domain.notion.dto.request.NotionCreatePageRequest;
+import pingpong.backend.domain.notion.dto.request.NotionPageUpdateRequest;
 import pingpong.backend.domain.notion.dto.response.DatabaseCreatedResponse;
 import pingpong.backend.domain.notion.dto.response.DatabaseWithPagesResponse;
 import pingpong.backend.domain.notion.dto.response.PageDetailResponse;
