+++ src/main/java/pingpong/backend/domain/notion/dto/request/NotionOAuthExchangeRequest.java
@@ -1,4 +1,4 @@
-package pingpong.backend.domain.notion.dto;
+package pingpong.backend.domain.notion.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
