+++ src/main/java/pingpong/backend/domain/notion/dto/response/NotionOAuthExchangeResponse.java
@@ -1,4 +1,4 @@
-package pingpong.backend.domain.notion.dto;
+package pingpong.backend.domain.notion.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
