+++ src/main/java/pingpong/backend/domain/notion/dto/response/NotionOAuthTokenResponse.java
@@ -1,4 +1,4 @@
-package pingpong.backend.domain.notion.dto;
+package pingpong.backend.domain.notion.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
