+++ src/main/java/pingpong/backend/domain/notion/service/NotionFacade.java
@@ -9,11 +9,11 @@ import pingpong.backend.domain.member.Member;
 import pingpong.backend.domain.notion.Notion;
 import pingpong.backend.domain.notion.NotionErrorCode;
 import pingpong.backend.domain.notion.client.NotionRestClient;
-import pingpong.backend.domain.notion.dto.NotionOAuthExchangeResponse;
-import pingpong.backend.domain.notion.dto.NotionOAuthTokenResponse;
+import pingpong.backend.domain.notion.dto.response.NotionOAuthExchangeResponse;
+import pingpong.backend.domain.notion.dto.response.NotionOAuthTokenResponse;
 import pingpong.backend.domain.notion.client.NotionOauthClient;
-import pingpong.backend.domain.notion.dto.NotionCreatePageRequest;
-import pingpong.backend.domain.notion.dto.NotionPageUpdateRequest;
+import pingpong.backend.domain.notion.dto.request.NotionCreatePageRequest;
+import pingpong.backend.domain.notion.dto.request.NotionPageUpdateRequest;
 import pingpong.backend.domain.notion.dto.response.DatabaseCreatedResponse;
 import pingpong.backend.domain.notion.dto.response.DatabaseWithPagesResponse;
 import pingpong.backend.domain.notion.dto.response.PageDetailResponse;
