+++ src/main/java/pingpong/backend/domain/notion/service/NotionPageService.java
@@ -10,11 +10,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import pingpong.backend.domain.notion.NotionErrorCode;
 import pingpong.backend.domain.notion.client.NotionRestClient;
-import pingpong.backend.domain.notion.dto.NotionCreatePageRequest;
-import pingpong.backend.domain.notion.dto.NotionDateRange;
-import pingpong.backend.domain.notion.dto.NotionPageUpdateRequest;
+import pingpong.backend.domain.notion.dto.request.NotionCreatePageRequest;
+import pingpong.backend.domain.notion.dto.common.NotionDateRange;
+import pingpong.backend.domain.notion.dto.request.NotionPageUpdateRequest;
 import pingpong.backend.domain.notion.dto.response.DatabaseWithPagesResponse;
-import pingpong.backend.domain.notion.dto.response.PageDateRange;
+import pingpong.backend.domain.notion.dto.common.PageDateRange;
 import pingpong.backend.domain.notion.dto.response.PageDetailResponse;
 import pingpong.backend.domain.notion.service.NotionPropertyResolver.PropertyNames;
 import pingpong.backend.domain.notion.util.NotionJsonUtils;
